gfhfgh

+++ BusinessObject/StudentRegistrationBO.cs
@@ -11,7 +11,7 @@ namespace BusinessObject
         private int id;
         private string firstName;
         private string lastName;
-        private int contactNumber;
+        private long contactNumber;
         private int collegeId;
         private DateTime createdDate;
 
@@ -19,7 +19,7 @@ namespace BusinessObject
         {
 
         }
-        public StudentRegistrationBO(int id, string firstName, string lastName, int contactNumber, int collegeId, DateTime createdDate)
+        public StudentRegistrationBO(int id, string firstName, string lastName, long contactNumber, int collegeId, DateTime createdDate)
         {
             this.id = id;
             this.firstName = firstName;
@@ -32,7 +32,7 @@ namespace BusinessObject
         public int Id { get => id; set => id = value; }
         public string FirstName { get => firstName; set => firstName = value; }
         public string LastName { get => lastName; set => lastName = value; }
-        public int ContactNumber { get => contactNumber; set => contactNumber = value; }
+        public long ContactNumber { get => contactNumber; set => contactNumber = value; }
         public int CollegeId { get => collegeId; set => collegeId = value; }
         public DateTime CreatedDate { get => createdDate; set => createdDate = value; }
     }
